Role Base Authentication is Done

+++ src/main/java/com/talib/journalApp/service/JournalEntryService.java
@@ -8,6 +8,7 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -24,12 +25,19 @@ public class JournalEntryService {//here we write our business logit
 
 
     //method to save entry and post Mapping
+    @Transactional//if anything is crash in this method then all the thing gonna rollback because lin 34 gonna crash so we can rollback all transection that happen before line 34
     public void saveEntry (JournalEntry JournalEntry,String userName) {
-        User user = userService.findByUsername(userName);
-        JournalEntry.setDate(LocalDateTime.now());
-        JournalEntry saved = JournalEntryRepository.save(JournalEntry);
-        user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+        try {
+            User user = userService.findByUsername(userName);
+            JournalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = JournalEntryRepository.save(JournalEntry);
+            user.getJournalEntries().add(saved);
+            //user.setUsername(null);
+            //userService.saveEntry(user);
+            userService.saveNewUser(user);
+        } catch (Exception e) {
+            throw new RuntimeException("en error occure while saving the entry",e);
+        }
     }
 
     public void saveEntry (JournalEntry JournalEntry) {
@@ -48,14 +56,27 @@ public class JournalEntryService {//here we write our business logit
         return JournalEntryRepository.findById(id);
     }
 
+    @Transactional
     //method for delete mapping
-    public void  deleteById(ObjectId id, String userName){
-        User user = userService.findByUsername(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        JournalEntryRepository.deleteById(id);
+    public boolean  deleteById(ObjectId id, String userName){
+        boolean removed=false;
+        try {
+            User user = userService.findByUsername(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed){
+                userService.saveEntry(user);
+                JournalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry.",e);
+        }return removed;
 
     }
+//    public List<JournalEntry> findByUserName(String username){
+//
+//
+//    }
 
 
 }
